naming, progress indicator, tests, task-player component adjustments
