+++ frontend/app/components/task-player/template.hbs
@@ -1,48 +1,6 @@
-<br />
-<AudioPlayer @audioFileUrl={{@task.audioFileUrl}} @autoplay={{true}} />
-{{#if (eq this.lastAnswer @task.word)}}
-  {{#if (not (and this.exerciseResultIsVisible @task.isLastExerciseTask))}}
-    <div
-      class="task-player__right-answer-data flex flex-wrap flex-1 mt-5"
-      {{did-insert (action this.runNextTaskTimer)}}
-    >
-      {{#if @task.pictureFileUrl}}
-        {{! template-lint-configure no-inline-styles false }}
-        <div
-          data-test-right-answer-notification
-          aria-label={{html-safe (concat "изображение " @task.word)}}
-          class="task-player__image flex-1 rounded shadow-md mx-auto"
-          role="img"
-        ></div>
-      {{/if}}
-    </div>
-  {{else}}
-    <AnswerCorrectnessWidget @isCorrect={{true}} />
-  {{/if}}
-{{else if (and @task.nextAttempt this.taskResultIsVisible)}}
-  <AnswerCorrectnessWidget
-    @isCorrect={{false}}
-    {{did-insert (action this.runNextAttemptTimer)}}
-  />
-{{else}}
-  <ul class="task-player__options mt-5">
-    {{#each this.shuffledWords as |word|}}
-      <li class="task-player__option">
-        <button
-          data-test-task-answer
-          data-test-task-answer-option={{word}}
-          disabled={{this.audio.player.isPlaying}}
-          onclick={{action this.handleSubmit word}}
-          class="task-player__option-button bg-transparent py-2 px-4 border border-blue-500 rounded text-blue-700
-            {{if
-              this.audio.player.isPlaying
-              "opacity-50 cursor-default"
-              "hover:bg-blue-500 hover:text-white hover:border-transparent"
-            }}"
-        >
-          {{word}}
-        </button>
-      </li>
-    {{/each}}
-  </ul>
-{{/if}}
+{{component
+  this.componentType
+  onRightAnswer=this.onRightAnswer
+  afterCompleted=this.afterCompleted
+  task=@task
+}}
