+++ frontend/app/components/task-player/words-sequences/template.hbs
