+++ frontend/app/helpers/sum.js
@@ -1,0 +1,5 @@
+import { helper } from '@ember/component/helper';
+
+export default helper(function sum([first, second]) {
+  return first + second;
+});
