+++ frontend/app/models/task/single-words.js
@@ -2,9 +2,10 @@ import DS from 'ember-data';
 const { attr } = DS;
 import BaseTask from '../task';
 
-export default class SingleWords extends BaseTask.extend({
+export default class SingleWordTask extends BaseTask.extend({
   word: attr('string'),
   audioFileUrl: attr('string'),
   pictureFileUrl: attr('string'),
   words: attr('array'),
+  exerciseType: 'single-words',
 }) {}
