+++ frontend/app/templates/series/exercise/task.hbs
@@ -1,10 +1,8 @@
-{{#if (eq this.model.exerciseType "SINGLE_WORDS")}}
-  <TaskPlayer
-    data-test-task-player
-    @onRightAnswer={{action this.saveExerciseMaybe}}
-    @afterCompleted={{action this.nextTaskTransition}}
-    @task={{this.model}}
-  />
-{{else if (eq this.model.exerciseType "WORDS_SEQUENCES")}}
-  <WordsSeqTaskPlayer @task={{this.model}} />
-{{/if}}
+<TaskPlayer
+  data-test-task-player
+  data-test-task-id={{this.model.id}}
+  data-test-task-exercise-id={{this.model.exercise.id}}
+  @onRightAnswer={{action this.saveExerciseMaybe}}
+  @afterCompleted={{action this.nextTaskTransition}}
+  @task={{this.model}}
+/>
